Fix the EmptyObjectAsConst canonicalizer rule
